Create analyzed to identify the peak centers
In this step of the research, data from an Excel file was imported and analyzed to identify the peak centers in specific ranges of angles. This analysis involved smoothing the data, identifying local maxima in the intensity values, and then extracting the corresponding peak centers. The resulting peak center data was saved to a .mat file. This process was automated using a MATLAB script, which can be found in the project repository on GitHub.